Remove msk dependency and move CreditCard and SavedCard to component folder
