+++ react/Payment.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import CreditCard from './CreditCard'
+import CreditCard from './components/CreditCard'
 
 const Payment: React.FC = () => {
   return <CreditCard />
